fix(equipment): address /review findings (>44)
- .shot import (importShotRecord) regression: the parsed grinder identity was
  silently dropped (migration 23 removed the snapshot columns and the INSERT was
  updated to match, but never set equipment_id). Now find-or-create an equipment
  package for the parsed identity and link equipment_id + rpm, mirroring the live
  shot-save path. New regression test v23_importShotRecordLinksEquipment.
- qsizetype: int loop counters over .size()/QJsonArray → qsizetype in new code
  (equipmentstorage.cpp insert/inventory loops, settings_network.cpp equipment
  layout-injection loop) per CLAUDE.md.
- createTables no longer issues CREATE INDEX idx_shots_grinder — it errored
  ("no such column") on every post-migration-23 launch and was silently swallowed.
- requestDeletePackage now also blocks hard-delete when another package's
  superseded_by points at the target, so a fork target can't be deleted out from
  under an older package's lineage pointer.
- activeEquipmentId widened int → qint64 (matches the qint64 ids used everywhere
  else: CoffeeBag/ShotRecord/ShotMetadata.equipmentId, EquipmentStorage ids).
- Stale migration-20 DECENZA_TESTING comment updated for migrations 22/23.

Full suite green (2395 passed, 0 failed).

+++ src/core/settings_dye.h
@@ -39,7 +39,7 @@ class SettingsDye : public QObject {
     Q_PROPERTY(int dyeGrinderRpm READ dyeGrinderRpm WRITE setDyeGrinderRpm NOTIFY dyeGrinderRpmChanged)
     // The active equipment package id (the grinder the next shot is ground on).
     // Switching it applies the package's grinder identity + last dial.
-    Q_PROPERTY(int activeEquipmentId READ activeEquipmentId WRITE setActiveEquipmentId NOTIFY activeEquipmentIdChanged)
+    Q_PROPERTY(qint64 activeEquipmentId READ activeEquipmentId WRITE setActiveEquipmentId NOTIFY activeEquipmentIdChanged)
     Q_PROPERTY(double dyeBeanWeight READ dyeBeanWeight WRITE setDyeBeanWeight NOTIFY dyeBeanWeightChanged)
     Q_PROPERTY(double dyeDrinkWeight READ dyeDrinkWeight WRITE setDyeDrinkWeight NOTIFY dyeDrinkWeightChanged)
     Q_PROPERTY(double dyeDrinkTds READ dyeDrinkTds WRITE setDyeDrinkTds NOTIFY dyeDrinkTdsChanged)
@@ -108,8 +108,8 @@ public:
     int dyeGrinderRpm() const;
     void setDyeGrinderRpm(int value);
 
-    int activeEquipmentId() const;
-    void setActiveEquipmentId(int id);
+    qint64 activeEquipmentId() const;
+    void setActiveEquipmentId(qint64 id);
 
     // User-facing equipment switch: adopt the package's grinder identity and its
     // last dial (grind setting + rpm), and point the active bag at it. `pkg` is a
